Fixed some bugs and added collectibles to the game that drop on monsters' deaths . I also fixed the problem with one monster being several megabytes by encapsulating all textures in shared_ptr to avoid copying shared textures.
The textures of the player going right and left changed for a better look but the animation was broken in the process (very likely to change).

Finally, a new circle was added 'collectible circle' which is basically the range for the player to collect items.

Next up is a sort of UI (the life bar on the top right) and also the collectibler count as well as money.
(And also going to add so that the player gets hurt when touched by a monster).

By the time I finish this I will have finished the class monster and I will thus be able to implement other types thereof to have some little more unique behaviors with some collisions between them (better than having monsters that constatnly follows you on a straight line and that can merge together).