Finished the Corpse Lurker, and added the Puffer to the game.
All monster types are now finished, so I'm moving on to the Bunker (to sell the flesh we have).

Thus, I'll need to make a new base class and change my tile system a bit. I'll make a new class that can bind to a tile and display something over that latter.

Thise class will allow the player to interact with it by pressing a key (while the mouse is on the tile bound to it).

This class will then call an Interface to actually display either the seller (bunker), wave calling system (campfire), or the shop (refugee camp).

When I say bind, it doesn't really bind to a tile with a pointer or reference to it (as objects of this base class will be placed in a vector). They will rather store the position of the tile. And since they must have a tile to exist they're technically "bound to it" even though what they ever truly have in common is a position.