Added places to the game, a Place binds to a position and a size.
It center itself in between the position and the size (size being the position plus some other factors).

Fixed some design issues with the Zone class too.
Now, at least one place is required for a map to exist and a place can be interacted with provided that the mouse is within its coordinates.